Files edited in another copy
The other copy (local) now works for 2019 libs,
at least enough not to notice a glaring error.

getline_m4_exp.cpp seems to be the only file
affected.

These files are a copy of the three edited in
the test repository, as they were when the
system compiled correctly.

They have not been added directly to the working
code in the present repository.

	new file:   common.h
	new file:   getline_m4_exp.cpp
	new file:   target_sensed.cpp

On branch testing-jj-borked-aa-